Add unit test
Added 1 more JUnit test and update code quality
